Changed the design of the chat image. Refactored and moved code to a new class. Added a library tha generates initials name image for users that doesn't have profile image.

+++ messaging/ShapeFormatterHelper.h
@@ -10,7 +10,7 @@
 
 @interface ShapeFormatterHelper : NSObject
 
-+(void)setRoundedView:(UIView *)roundedView toDiameter:(float)newSize;
++(void)setRoundedView:(UIView *)roundedView toDiameter:(CGFloat)newSize;
 
 + (void)setRoundedViewWithNotClipToBounds:(UIView *)roundedView toDiameter:(float)newSize;
